backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv
slowapi
PyJWT
email-validator